+++ idb_sqflite/lib/src/idb_import.dart
@@ -1,5 +1,6 @@
 export 'package:idb_shim/src/common/common_database.dart';
 export 'package:idb_shim/src/common/common_exception.dart';
+export 'package:idb_shim/src/common/common_join_query.dart';
 export 'package:idb_shim/src/common/common_meta.dart';
 export 'package:idb_shim/src/common/common_paged_query.dart';
 export 'package:idb_shim/src/common/common_transaction.dart';